App_FPGAging_Slack_Sensor/requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
ifaddr==0.2.0
kiwisolver==1.4.9
numpy==2.3.5
packaging==25.0
pillow==12.0.0
psutil==7.2.1
pyparsing==3.2.5
pyqtgraph==0.14.0
pyserial==3.5
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
python-usbtmc==0.8
pyusb==1.3.1
PyVISA==1.16.0
PyVISA-py==0.8.1
shiboken6==6.10.0
six==1.17.0
typing_extensions==4.15.0
zeroconf==0.148.0
